break the errors classes into multiple files for JSDoc rendering
